can't get json to come up. Installed Json add on but will not come up with the options.
Working to see if weight and quantity of water could be added to screen. Tried putting them all in Mainpage and both worked but when put weight back, weight gets into field value of MeasurementSystem but then when try to call DrinkAmount(), field value is back at 0.
Probs resets bc going back to Mainpage to call it